# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeeklyRun API",
    "description": "WeeklyRun - AI Meeting Orchestration API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/minutes/meetings/{meeting_id}/generate-minutes": {
      "post": {
        "tags": [
          "minutes"
        ],
        "summary": "Start Minutes Generation",
        "description": "Start meeting minutes generation.\n\nRequires transcription to be completed first.\nUses GPT to generate minutes from transcript and weekly report.",
        "operationId": "start_minutes_generation_api_v1_minutes_meetings__meeting_id__generate_minutes_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinutesGenerationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/minutes/meetings/{meeting_id}/minutes": {
      "get": {
        "tags": [
          "minutes"
        ],
        "summary": "Get Meeting Minutes",
        "description": "Get meeting minutes.",
        "operationId": "get_meeting_minutes_api_v1_minutes_meetings__meeting_id__minutes_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinutesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "minutes"
        ],
        "summary": "Update Meeting Minutes",
        "description": "Update meeting minutes content.\n\nStores the edited version and marks as edited.",
        "operationId": "update_meeting_minutes_api_v1_minutes_meetings__meeting_id__minutes_put",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MinutesUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinutesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/minutes/meetings/{meeting_id}/publish": {
      "post": {
        "tags": [
          "minutes"
        ],
        "summary": "Publish Minutes To Confluence",
        "description": "Publish meeting minutes to Confluence.\n\nCreates a new Confluence page with the minutes content.",
        "operationId": "publish_minutes_to_confluence_api_v1_minutes_meetings__meeting_id__publish_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/minutes/meetings/{meeting_id}/minutes/export": {
      "get": {
        "tags": [
          "minutes"
        ],
        "summary": "Export Minutes",
        "description": "Export meeting minutes in various formats.\n\nSupported formats: markdown, html",
        "operationId": "export_minutes_api_v1_minutes_meetings__meeting_id__minutes_export_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "markdown",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Minutes Api V1 Minutes Meetings  Meeting Id  Minutes Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorrectionItemResponse": {
        "properties": {
          "original": {
            "type": "string",
            "title": "Original"
          },
          "corrected": {
            "type": "string",
            "title": "Corrected"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "paragraph_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph Index"
          },
          "start_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Offset"
          },
          "end_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Offset"
          }
        },
        "type": "object",
        "required": [
          "original",
          "corrected",
          "category"
        ],
        "title": "CorrectionItemResponse",
        "description": "Single correction item."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MinutesGenerationStatusResponse": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "has_minutes": {
            "type": "boolean",
            "title": "Has Minutes"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "status",
          "has_minutes",
          "error_message"
        ],
        "title": "MinutesGenerationStatusResponse",
        "description": "Status response for minutes generation."
      },
      "MinutesResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "content_markdown": {
            "type": "string",
            "title": "Content Markdown"
          },
          "ai_model": {
            "type": "string",
            "title": "Ai Model"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "is_edited": {
            "type": "boolean",
            "title": "Is Edited"
          },
          "edited_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Content"
          },
          "corrections": {
            "items": {
              "$ref": "#/components/schemas/CorrectionItemResponse"
            },
            "type": "array",
            "title": "Corrections"
          }
        },
        "type": "object",
        "required": [
          "id",
          "meeting_id",
          "content_markdown",
          "ai_model",
          "prompt_version",
          "is_edited",
          "edited_content",
          "corrections"
        ],
        "title": "MinutesResponse",
        "description": "Meeting minutes response."
      },
      "MinutesUpdateRequest": {
        "properties": {
          "content_markdown": {
            "type": "string",
            "title": "Content Markdown"
          }
        },
        "type": "object",
        "required": [
          "content_markdown"
        ],
        "title": "MinutesUpdateRequest",
        "description": "Request to update minutes content."
      },
      "PublishResponse": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "confluence_page_id": {
            "type": "string",
            "title": "Confluence Page Id"
          },
          "confluence_page_url": {
            "type": "string",
            "title": "Confluence Page Url"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "confluence_page_id",
          "confluence_page_url"
        ],
        "title": "PublishResponse",
        "description": "Response after publishing to Confluence."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}